got factory working
Factory is finally working. Was a minor bug, I was checking if the
Factory had reached its limit but checking if the number alive is
**greater than** the limit. Which is stupid. But it works now.

Currently, monsters and woman NPCs are generated at random. I need to
create spawning points for all of them. This will be added in the new
map.